requirements.txt
ansible==2.10.3
ansible-base==2.10.3
ansible-inventory==0.5.20
ansible-netbox-inventory==1.0.9
assertpy==1.1
attrs==20.2.0
autopep8==1.5.4
backcall==0.2.0
bcrypt==3.2.0
bigsuds==1.0.6
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
ciscoconfparse==1.5.19
click==7.1.2
colorama==0.4.4
commonmark==0.9.1
configparser==5.0.1
cryptography==3.4.6
DateTime==4.3
decorator==4.4.2
decorators==2.0.7
dnspython==2.0.0
et-xmlfile==1.0.1
Exscript==2.6.3
f5-icontrol-rest==1.3.13
f5-sdk==3.0.21
fmcapi==20210203.0
future==0.18.2
graphviz==0.16
idna==2.10
iniconfig==1.1.1
ipaddress==1.0.23
ipython==7.18.1
ipython-genutils==0.2.0
jdcal==1.4.1
jedi==0.17.2
Jinja2==2.11.3
jsonpath-ng==1.5.2
junos-eznc==2.5.4
lxml==4.6.1
MarkupSafe==1.1.1
mypy-extensions==0.4.3
napalm==3.2.0
ncclient==0.6.9
netaddr==0.8.0
netmiko==3.3.2
nmap==0.0.1
nornir==3.0.0
nornir-jinja2==0.1.0
nornir-napalm==0.1.1
nornir-netbox==0.1.1
nornir-netmiko==0.1.1
nornir-tests==0.1.0a11
nornir-utils==0.1.1
ntc-ansible==0.9.2
ntc-templates==1.6.0
openpyxl==3.0.5
packaging==20.4
paramiko==2.7.2
parso==0.7.1
passlib==1.7.4
pathlib==1.0.1
pexpect==4.8.0
pickleshare==0.7.5
pip-tools==5.3.1
pkg-resources==0.0.0
pluggy==0.13.1
ply==3.11
prompt-toolkit==3.0.8
ptyprocess==0.6.0
py==1.9.0
pyaml==20.4.0
pycodestyle==2.6.0
pycparser==2.20
pycryptodomex==3.9.9
pyeapi==0.8.3
pyexpect==1.0.20
Pygments==2.7.1
PyNaCl==1.4.0
pyntc==0.0.13
pynxos==0.0.5
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyserial==3.4
pyspellchecker==0.6.2
pytest==6.1.1
pytest-clarity==0.3.0a0
pytz==2021.1
PyYAML==5.3.1
redis==3.5.3
requests==2.24.0
rich==7.1.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
scp==0.13.2
six==1.15.0
suds-jurko==0.6
tabulate==0.8.7
tenacity==6.2.0
termcolor==1.1.0
textfsm==1.1.0
toml==0.10.1
traitlets==5.0.5
transitions==0.8.5
typing-extensions==3.7.4.3
urllib3==1.25.11
wcwidth==0.2.5
wrapt==1.12.1
xlrd==1.2.0
yamlordereddictloader==0.4.0
zope.interface==5.2.0
